requirements.txt
bandit==1.7.5
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.1.3
flake8==6.1.0
Flask==3.0.0
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
oauthlib==3.2.2
packaging==23.2
pbr==5.11.1
pluggy==1.3.0
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.16.1
pytest==7.4.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
smmap==5.0.1
stevedore==5.1.0
tomli==2.0.1
tweepy==4.14.0
urllib3==2.0.7
Werkzeug==3.0.0
